01-mcp-study/1_basic/pyproject.toml
[project]
name = "mcp-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ddgs>=9.5.0",
    "google-genai>=1.28.0",
    "mcp[cli]>=1.12.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
